implement a `PipelineConfiguration` test (Pipeline.from_yaml versus Pipeline.from_config)

+++ src/biome/text/features.py
@@ -59,7 +59,7 @@ class CharFeatures:
         self,
         embedding_dim: int,
         encoder: Dict[str, Any],
-        dropout: int = 0.0,
+        dropout: float = 0.0,
         **extra_params
     ):
         self.embedding_dim = embedding_dim
